Fixed GuiTrack contructor (issue 272) http://code.google.com/p/jmonkeyengine/issues/detail?id=272&q=status:New&sort=-product&colspec=ID%20Type%20Status%20Component%20Priority%20Product%20Milestone%20Owner%20Summary#makechanges

+++ branches/jme3/src/core/com/jme3/cinematic/events/GuiTrack.java
@@ -68,7 +68,7 @@ public class GuiTrack extends AbstractCinematicEvent {
         this.screen = screen;
     }
 
-    public GuiTrack(String screen, int initialDuration, LoopMode loopMode) {
+    public GuiTrack(String screen, float initialDuration, LoopMode loopMode) {
         super(initialDuration, loopMode);
         this.screen = screen;
     }
